backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
paho-mqtt==1.6.1
